pyproject.toml
[project]
name = "accprob"
version = "0.1.0"
description = "AtCoder Problems command line tools"
authors = [
    { name = "take0x" },
    { name = "Hietan" },
    { name = "n4okins" },
]
dependencies = [
    "selenium>=4.20.0",
    "online-judge-tools>=11.5.1",
    "setuptools>=69.5.1",
    "pydantic>=2.7.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
accprob = "accprob.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "mypy>=1.10.0",
    "ruff>=0.4.2",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/accprob"]

[tool.mypy]
strict = true

[tool.ruff]
target-version = "py311"
lint.select = ["ALL"]
lint.ignore = [
    "D",       # pydocstyle
    "ANN1",    # Missing type annotation for 'self' and 'cls'
    "S603",    # 'subprocess' call: check for execution of untrusted input
    "COM812",  # Trailing comma missing
    "ISC001",  # Implicitly concatenated string literals on one line
    "T201",    # 'print' found
]
